requirements.txt
teslapy>=2.8.0
requests>=2.28.0
rumps>=0.4.0
python-dotenv>=1.0.0
